requirements.txt
Flask==3.0.3
jinja2==3.1.4
python-dotenv==1.0.1
